move MessageRecyclerItem
